requirements.txt
bedrock-agentcore
strands-agents
pandas
matplotlib
reportlab
seaborn
numpy